Refactor Molmo2: merge into existing modules, drop TF5 patches, pin transformers<5
Changes:
- Merge Molmo2Template into swift/template/templates/molmo.py (was standalone molmo2.py)
- Merge Molmo2Loader into swift/model/models/mllm.py (was standalone molmo2.py)
- Delete standalone swift/template/templates/molmo2.py and swift/model/models/molmo2.py
- Remove __init__.py imports of deleted molmo2 modules
- Remove all transformers>=5.0 compatibility patches:
  - ProcessorMixin.__init__ monkey-patch for unknown kwargs
  - ROPE_INIT_FUNCTIONS['default'] injection
  - prepare_inputs_for_generation cache_position fix
- Pin version to transformers>=4.57.1,<5 (verified working)
- Simplify Molmo2Loader from ~80 lines to ~10 lines

Template (Molmo2Template):
- ChatML format via ChatmlTemplateMeta with auto_add_bos=True
- replace_tag returns '<|image|>' / '<|video|>' placeholders
- _prepare_mm_inputs: processor.image_processor for images,
  processor.video_processor for videos (with return_metadata=True)
- _encode: super()._encode() + _extend_tokens to expand placeholders
- _build_token_type_ids: marks image_token_ids as 1
- _data_collator_mm_data: handles image_grids, video_grids, etc.

Model (Molmo2Loader):
- AutoModelForImageTextToText as auto_model_cls
- get_class_from_dynamic_module for trust_remote_code
- _no_split_modules += MolmoSequentialBlock
- patch_output_clone on wte embedding

Tested on 4x H100 with transformers==4.57.6, torch==2.8.0:
- Image inference (TransformersEngine): PASS
- Video inference (TransformersEngine): PASS
- LoRA SFT training (4 steps, rank=8): PASS
